+++ gbe/views/act_display_functions.py
@@ -1,56 +1,11 @@
-from django.contrib import messages
-from django.urls import reverse
-from django.shortcuts import render
 from django.forms import (
     ChoiceField,
-)
-from gbe.models import (
-    Act,
-    ActCastingOption,
-    UserMessage,
 )
 from gbetext import (
     act_shows_options,
     old_act_shows_options,
-    default_act_title_conflict,
-    act_not_unique,
 )
 from gbe_forms_text import act_bid_labels
-
-
-def display_invalid_act(request, data, form, conference, profile, view):
-    if [act_not_unique] in list(form.errors.values()):
-        conflict_msg = UserMessage.objects.get_or_create(
-            view=view,
-            code="ACT_TITLE_CONFLICT",
-            defaults={
-                'summary': "Act Title, User, Conference Conflict",
-                'description': default_act_title_conflict})
-        conflict = Act.objects.filter(
-            b_conference=conference,
-            b_title=form.data['theact-b_title'],
-            bio__contact=profile).first()
-        if conflict.submitted:
-            link = reverse(
-                'act_view',
-                urlconf='gbe.urls',
-                args=[conflict.pk]
-            )
-        else:
-            link = reverse(
-                'act_edit',
-                urlconf='gbe.urls',
-                args=[conflict.pk]
-            )
-        messages.error(
-            request, conflict_msg[0].description % (
-                link,
-                conflict.b_title))
-    return render(
-        request,
-        'gbe/bid.tmpl',
-        data
-    )
 
 
 def get_act_form(act, form, header):
@@ -74,13 +29,3 @@ def get_act_form(act, form, header):
         act_shows_options
 
     return act_form
-
-
-def get_act_casting():
-    castings = ActCastingOption.objects.all()
-    cast_list = []
-    for casting in castings:
-        value = casting.casting
-        cast_list += [(value, casting.casting)]
-
-    return cast_list
